+++ src/Controller/DefaultController.php
@@ -48,7 +48,7 @@ final class DefaultController extends BaseController
     {
         $userService = $this->container->get('user_service');
         $taskService = $this->container->get('task_service');
-        $noteService = $this->container->get('note_service');
+        $noteService = $this->container->get('get_note_service');
 
         return [
             'users' => count($userService->getAll()),
