+++ src/Controller/Note/GetAll.php
@@ -11,7 +11,7 @@ final class GetAll extends Base
 {
     public function __invoke(Request $request, Response $response): Response
     {
-        $notes = $this->getNoteService()->getAll();
+        $notes = $this->getServiceGetNote()->getAll();
 
         return $this->jsonResponse($response, 'success', $notes, 200);
     }
