+++ src/Controller/Note/GetOne.php
@@ -11,7 +11,7 @@ final class GetOne extends Base
 {
     public function __invoke(Request $request, Response $response, array $args): Response
     {
-        $note = $this->getNoteService()->getOne((int) $args['id']);
+        $note = $this->getServiceGetNote()->getOne((int) $args['id']);
 
         return $this->jsonResponse($response, 'success', $note, 200);
     }
