+++ src/Controller/Note/Search.php
@@ -11,7 +11,7 @@ final class Search extends Base
 {
     public function __invoke(Request $request, Response $response, array $args): Response
     {
-        $notes = $this->getNoteService()->search($args['query']);
+        $notes = $this->getServiceGetNote()->search($args['query']);
 
         return $this->jsonResponse($response, 'success', $notes, 200);
     }
