+++ src/Service/Note/NoteService.php
@@ -6,24 +6,4 @@ namespace App\Service\Note;
 
 final class NoteService extends Base
 {
-    public function getAll(): array
-    {
-        return $this->noteRepository->getNotes();
-    }
-
-    public function getOne(int $noteId): object
-    {
-        if (self::isRedisEnabled() === true) {
-            $note = $this->getOneFromCache($noteId);
-        } else {
-            $note = $this->getOneFromDb($noteId);
-        }
-
-        return $note;
-    }
-
-    public function search(string $notesName): array
-    {
-        return $this->noteRepository->searchNotes($notesName);
-    }
 }
